+++ lib/nokogiri/html4/document.rb
@@ -159,7 +159,6 @@ module Nokogiri
         # Nokogiri::XML::ParseOptions.
         def parse(string_or_io, url = nil, encoding = nil, options = XML::ParseOptions::DEFAULT_HTML)
           options = Nokogiri::XML::ParseOptions.new(options) if Integer === options
-
           yield options if block_given?
 
           url ||= string_or_io.respond_to?(:path) ? string_or_io.path : nil