+++ lib/nokogiri/html4/document_fragment.rb
@@ -5,7 +5,7 @@ module Nokogiri
     class DocumentFragment < Nokogiri::XML::DocumentFragment
       ####
       # Create a Nokogiri::XML::DocumentFragment from +tags+, using +encoding+
-      def self.parse(tags, encoding = nil)
+      def self.parse(tags, encoding = nil, options = XML::ParseOptions::DEFAULT_HTML, &block)
         doc = HTML4::Document.new
 
         encoding ||= if tags.respond_to?(:encoding)
@@ -21,15 +21,18 @@ module Nokogiri
 
         doc.encoding = encoding
 
-        new(doc, tags)
+        new(doc, tags, nil, options, &block)
       end
 
-      def initialize(document, tags = nil, ctx = nil)
+      def initialize(document, tags = nil, ctx = nil, options = XML::ParseOptions::DEFAULT_HTML)
         return self unless tags
+
+        options = Nokogiri::XML::ParseOptions.new(options) if Integer === options
+        yield options if block_given?
 
         if ctx
           preexisting_errors = document.errors.dup
-          node_set = ctx.parse("<div>#{tags}</div>")
+          node_set = ctx.parse("<div>#{tags}</div>", options)
           node_set.first.children.each { |child| child.parent = self } unless node_set.empty?
           self.errors = document.errors - preexisting_errors
         else
@@ -40,7 +43,7 @@ module Nokogiri
             "/html/body/node()"
           end
 
-          temp_doc = HTML4::Document.parse("<html><body>#{tags}", nil, document.encoding)
+          temp_doc = HTML4::Document.parse("<html><body>#{tags}", nil, document.encoding, options)
           temp_doc.xpath(path).each { |child| child.parent = self }
           self.errors = temp_doc.errors
         end